requirements.txt
# Core scientific dependencies
astropy==7.0.1
astropy-iers-data==0.2025.2.3.0.32.42
astroquery==0.4.9.post1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
filterpy==1.4.5
fonttools==4.56.0
geographiclib==2.0
geopy==2.4.1
html5lib==1.1
humanize==4.11.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
joblib==1.4.2
jplephem==2.22
keyring==25.6.0
kiwisolver==1.4.8
matplotlib==3.10.0
more-itertools==10.6.0
narwhals==1.26.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
pyerfa==2.0.1.5
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyvo==1.6
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sgp4==2.23
six==1.17.0
skyfield==1.49
soupsieve==2.6
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1

# API Framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP & Networking
httpx>=0.25.0
aiohttp>=3.9.0

# Machine Learning (additional)
torch>=2.1.0

# Monitoring & Metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# Caching
redis>=5.0.0

# Logging & Configuration
structlog>=23.2.0
python-dotenv>=1.0.0

# Templates & Static Files
jinja2>=3.1.0
aiofiles>=23.2.0

# Streaming
sse-starlette>=1.6.0
websockets>=12.0

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development Tools
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0
